added basic auth functionality
