+++ app/api/auth/callback/route.ts
@@ -16,9 +16,9 @@ export async function GET(request: Request) {
 			const isLocalEnv = process.env.NODE_ENV === "development";
 			if (isLocalEnv) {
 				// we can be sure that there is no load balancer in between, so no need to watch for X-Forwarded-Host
-				return NextResponse.redirect(`${origin}${next}`);
+				return NextResponse.redirect(`${origin}${next}/home`);
 			} else if (forwardedHost) {
-				return NextResponse.redirect(`https://${forwardedHost}${next}`);
+				return NextResponse.redirect(`https://${forwardedHost}${next}/home`);
 			} else {
 				return NextResponse.redirect(`${origin}${next}`);
 			}
