+++ app/home/page.tsx
@@ -1,8 +1,3 @@
 export default function Home() {
-	return (
-		<>
-			<h1>You are at the Home page</h1>
-			<h2>I want to test the vercel deployment </h2>
-		</>
-	);
+	return <p>Im da home page</p>;
 }
