+++ lib/utils/supabase/middleware.ts
@@ -39,7 +39,7 @@ export async function updateSession(request: NextRequest) {
 
 	if (
 		!user &&
-		!request.nextUrl.pathname.startsWith("/login") &&
+		!request.nextUrl.pathname.startsWith("/") &&
 		!request.nextUrl.pathname.startsWith("/auth")
 	) {
 		// no user, potentially respond by redirecting the user to the login page
